backend/src/v2/requirements.txt
uvicorn
fastapi
pydantic_settings
slowapi
requests
python-multipart
pillow
pyjwt
sqlalchemy
passlib[bcrypt]
pycryptodome
openai
httpx
num2words
pytz
pdf2image
fpdf2
chromadb
sentence-transformers
rank-bm25
pytest 
pytest-asyncio 
httpx
respx
sqlalchemy[asyncio] 
testcontainers